Return value type 'int' of method send() changed to 'size_t'.

+++ socket.h
@@ -23,13 +23,11 @@ public:
     Socket& operator = (Socket&& other);
 
     size_t bind(const std::string &address);
-//    int bind(const std::string &address);
     int connect(const std::string &address);
 
     void bind(const sockaddr_in &addr, size_t addrlen);
-//    int bind(const sockaddr_in &addr, size_t addrlen);
     void listen(int n);
-    int send(const void *buf, size_t len, int n);
+    size_t send(const void *buf, size_t len, int n);
     int recv(void *buf, size_t len, int n);
     int connect(const sockaddr_in &addr, size_t addrlen);
     Socket accept(sockaddr_in &addr, socklen_t &addrlen);
